move output tests into packages
Previously it was necessary to enter the "examples" module to run output tests
for code in the main module. Now "go test ./..." at the root or in individual
directories also runs these tests.
